+++ src/test/java/io/github/astrapi69/crypto/key/reader/PemObjectReaderTest.java
@@ -101,13 +101,15 @@ public class PemObjectReaderTest
 		File newCsrCertFile;
 		File pkcs7File;
 
-		// new scenario...
+			// new scenario...
 		rsaPrivatekeyPemFile = new File(pemDir, "private.pem");
 		assertTrue(PemObjectReader.isPemObject(rsaPrivatekeyPemFile));
 		pemObject = PemObjectReader.getPemObject(rsaPrivatekeyPemFile);
 		actual = pemObject.getType();
 		expected = PemType.RSA_PRIVATE_KEY_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.RSA_PRIVATE_KEY, PemObjectReader.getPemType(pemObject));
+
 		// new scenario...
 		privatekey2PemFile = new File(pemDir, "private2.pem");
 		assertTrue(PemObjectReader.isPemObject(privatekey2PemFile));
@@ -115,60 +117,70 @@ public class PemObjectReaderTest
 		actual = pemObject.getType();
 		expected = PemType.PRIVATE_KEY_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.PRIVATE_KEY, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		rsaPublickeyFile = new File(pemDir, "rsa-public-key.pem");
 		pemObject = PemObjectReader.getPemObject(rsaPublickeyFile);
 		actual = pemObject.getType();
 		expected = PemType.RSA_PUBLIC_KEY_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.RSA_PUBLIC_KEY, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		publickeyFile = new File(pemDir, "public.pem");
 		pemObject = PemObjectReader.getPemObject(publickeyFile);
 		actual = pemObject.getType();
 		expected = PemType.PUBLIC_KEY_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.PUBLIC_KEY, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		dsaPPPrivatekeyFile = new File(pemDir, "dsa-pwp-pk-pw-is-123456.pem");
 		pemObject = PemObjectReader.getPemObject(dsaPPPrivatekeyFile);
 		actual = pemObject.getType();
 		expected = PemType.DSA_PRIVATE_KEY_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.DSA_PRIVATE_KEY, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		rsaPPPrivatekeyFile = new File(pemDir, "rsa-pwp-pk-pw-is-123456.pem");
 		pemObject = PemObjectReader.getPemObject(rsaPPPrivatekeyFile);
 		actual = pemObject.getType();
 		expected = PemType.RSA_PRIVATE_KEY_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.RSA_PRIVATE_KEY, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		crlCertFile = new File(pemDir, "crl-cert.pem");
 		pemObject = PemObjectReader.getPemObject(crlCertFile);
 		actual = pemObject.getType();
 		expected = PemType.X509_CRL_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.X509_CRL, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		crtCertFile = new File(pemDir, "certificate.pem");
 		pemObject = PemObjectReader.getPemObject(crtCertFile);
 		actual = pemObject.getType();
 		expected = PemType.CERTIFICATE_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.CERTIFICATE, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		csrCertFile = new File(pemDir, "csr-cert.pem");
 		pemObject = PemObjectReader.getPemObject(csrCertFile);
 		actual = pemObject.getType();
 		expected = PemType.CERTIFICATE_REQUEST_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.CERTIFICATE_REQUEST, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		newCsrCertFile = new File(pemDir, "new-csr-cert.pem");
 		pemObject = PemObjectReader.getPemObject(newCsrCertFile);
 		actual = pemObject.getType();
 		expected = PemType.NEW_CERTIFICATE_REQUEST_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.NEW_CERTIFICATE_REQUEST, PemObjectReader.getPemType(pemObject));
 		// new scenario...
 		pkcs7File = new File(pemDir, "pkcs7.pem");
 		pemObject = PemObjectReader.getPemObject(pkcs7File);
 		actual = pemObject.getType();
 		expected = PemType.PKCS7_KEY_NAME;
 		assertEquals(expected, actual);
+		assertEquals(PemType.PKCS7_KEY, PemObjectReader.getPemType(pemObject));
 	}
 
 	/**